ASABASE-953 Device Emulator fixes
Fixed bugs in models
Fixed a empty buffer bug in HAL
Fixed a cast warning

+++ src/utils/CDM_AboutData.c
@@ -345,7 +345,7 @@ static int WriteDeviceTypesDescriptions(const ABOUT_ANALYSIS *analysis, void *de
         memcpy(curPos, analysis->devicePaths.values[i].element, analysis->devicePaths.values[i].elemSize);
         aboutDataBuf->itr += analysis->devicePaths.values[i].elemSize + 1;
 
-        desc[i].type = (DeviceType)atoi(analysis->deviceTypes.values[i].element);
+        desc[i].type = (DeviceType)(int)atoi(analysis->deviceTypes.values[i].element);
         desc[i].objectpath = curPos;
     }
